# tenant_platform/contracts/openapi/platform.yaml
openapi: 3.1.0
info:
  title: GenericAgent Tenant Platform Foundation API
  version: 0.1.0
  description: >
    Loopback foundation HTTP contract for session tasks. X-Platform-Admin-Token is
    required for this foundation-only API. No browser chat endpoint is exposed;
    the P0 user chat path remains BotTransportAdapter.
servers:
  - url: http://127.0.0.1:8080
    description: Local loopback foundation backend
security:
  - PlatformAdminToken: []
paths:
...
  /v1/admin/users:
    post:
      operationId: adminCreateUser
      summary: Create a pending platform user (admin-only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRequest"
      responses:
        "201":
          description: User created in pending state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRecord"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/admin/users/{user_id}/approve:
    post:
      operationId: adminApproveUser
      summary: Approve a pending user (pending → approved)
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRecord"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/admin/users/{user_id}/block:
    post:
      operationId: adminBlockUser
      summary: Block a user (approved → blocked); cancels queued/running tasks
      parameters:
        - $ref: "#/components/parameters/UserId"
      responses:
        "200":
          description: User blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserRecord"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/admin/users/pending:
    get:
      operationId: adminListPendingUsers
      summary: List users in pending state
      responses:
        "200":
          description: Pending users
          content:
            application/json:
              schema:
                type: object
                required:
                  - users
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/UserRecord"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/admin/llm-providers:
    get:
      operationId: adminListLLMProviders
      summary: List all LLM providers (admin-only)
      tags: [admin]
      security: [{PlatformAdminToken: []}]
      responses:
        "200":
          description: Provider list
          content:
            application/json:
              schema:
                type: object
                required:
                  - providers
                properties:
                  providers:
                    type: array
                    items: {$ref: "#/components/schemas/LLMProvider"}
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      operationId: adminCreateLLMProvider
      summary: Create an upstream LLM provider (admin-only)
      tags: [admin]
      security: [{PlatformAdminToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLLMProviderRequest"
      responses:
        "201":
          description: Provider created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LLMProvider"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/admin/llm-providers/{provider_id}:
    get:
      operationId: adminGetLLMProvider
      summary: Get an LLM provider by ID (admin-only)
      tags: [admin]
      security: [{PlatformAdminToken: []}]
      parameters:
        - name: provider_id
          in: path
          required: true
          schema: {type: integer, format: int64}
      responses:
        "200":
          description: Provider record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LLMProvider"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    put:
      operationId: adminUpdateLLMProvider
      summary: Update an LLM provider (admin-only)
      tags: [admin]
      security: [{PlatformAdminToken: []}]
      parameters:
        - name: provider_id
          in: path
          required: true
          schema: {type: integer, format: int64}
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateLLMProviderRequest"
      responses:
        "200":
          description: Updated provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LLMProvider"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
    delete:
      operationId: adminDeleteLLMProvider
      summary: Delete an LLM provider (admin-only)
      tags: [admin]
      security: [{PlatformAdminToken: []}]
      parameters:
        - name: provider_id
          in: path
          required: true
          schema: {type: integer, format: int64}
      responses:
        "204":
          description: Provider deleted
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/admin/llm-providers/{provider_id}/default:
    post:
      operationId: adminSetDefaultLLMProvider
      summary: Set the default LLM provider (admin-only)
      tags: [admin]
      security: [{PlatformAdminToken: []}]
      parameters:
        - name: provider_id
          in: path
          required: true
          schema: {type: integer, format: int64}
      responses:
        "200":
          description: Default provider set
          content:
            application/json:
              schema:
                type: object
                required:
                  - provider_id
                  - is_default
                properties:
                  provider_id:
                    type: integer
                    format: int64
                  is_default:
                    type: boolean
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/admin/llm-providers/{provider_id}/disable:
    post:
      operationId: adminDisableLLMProvider
      summary: Disable a non-default LLM provider (admin-only)
      tags: [admin]
      security: [{PlatformAdminToken: []}]
      parameters:
        - name: provider_id
          in: path
          required: true
          schema: {type: integer, format: int64}
      responses:
        "200":
          description: Disabled provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LLMProvider"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/admin/llm-providers/{provider_id}/enable:
    post:
      operationId: adminEnableLLMProvider
      summary: Enable an LLM provider (admin-only)
      tags: [admin]
      security: [{PlatformAdminToken: []}]
      parameters:
        - name: provider_id
          in: path
          required: true
          schema: {type: integer, format: int64}
      responses:
        "200":
          description: Enabled provider
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LLMProvider"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/Conflict"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/admin/mcp-servers:
    get:
      operationId: adminListMCPServers
      summary: List globally shared MCP servers (admin-only)
      tags: [admin]
      security: [{PlatformAdminToken: []}]
      responses:
        "200":
          description: MCP server list
          content:
            application/json:
              schema:
                type: object
                required: [servers]
                properties:
                  servers:
                    type: array
                    items: {$ref: "#/components/schemas/MCPServer"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "500": {$ref: "#/components/responses/InternalError"}
    post:
      operationId: adminCreateMCPServer
      summary: Create a disabled global MCP server (admin-only)
      tags: [admin]
      security: [{PlatformAdminToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/CreateMCPServerRequest"}
      responses:
        "201":
          description: Disabled MCP server created
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MCPServer"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "409": {$ref: "#/components/responses/Conflict"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/admin/mcp-servers/{mcp_server_id}:
    parameters:
      - name: mcp_server_id
        in: path
        required: true
        schema: {type: integer, format: int64, minimum: 1}
    put:
      operationId: adminUpdateMCPServer
      summary: Update a global MCP server (admin-only)
      tags: [admin]
      security: [{PlatformAdminToken: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UpdateMCPServerRequest"}
      responses:
        "200":
          description: MCP server updated
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MCPServer"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "409": {$ref: "#/components/responses/Conflict"}
        "500": {$ref: "#/components/responses/InternalError"}
    delete:
      operationId: adminDeleteMCPServer
      summary: Delete a global MCP server (admin-only)
      tags: [admin]
      security: [{PlatformAdminToken: []}]
      responses:
        "204": {description: MCP server deleted}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/admin/mcp-servers/{mcp_server_id}/enable:
    post:
      operationId: adminEnableMCPServer
      summary: Enable an MCP server for every tenant (admin-only)
      tags: [admin]
      security: [{PlatformAdminToken: []}]
      parameters:
        - name: mcp_server_id
          in: path
          required: true
          schema: {type: integer, format: int64, minimum: 1}
      responses:
        "200":
          description: MCP server enabled
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MCPServer"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/admin/mcp-servers/{mcp_server_id}/disable:
    post:
      operationId: adminDisableMCPServer
      summary: Disable an MCP server for every tenant (admin-only)
      tags: [admin]
      security: [{PlatformAdminToken: []}]
      parameters:
        - name: mcp_server_id
          in: path
          required: true
          schema: {type: integer, format: int64, minimum: 1}
      responses:
        "200":
          description: MCP server disabled
          content:
            application/json:
              schema: {$ref: "#/components/schemas/MCPServer"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "404": {$ref: "#/components/responses/NotFound"}
        "500": {$ref: "#/components/responses/InternalError"}
  /v1/router/messages:
    post:
      security:
        - PlatformAdminToken: []
      operationId: routerMessage
      summary: Ingest one bot message through the router pipeline
      description: >
        Processes identity resolution, user status check, command parsing
        (/stop, /new, /activate), and task routing. Replies are sent via
        the BotTransportAdapter; the HTTP response carries the action
        classification for observability.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouterMessage"
      responses:
        "200":
          description: Message processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RouterResult"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /v1/admin/commands:
    get:
      tags: [admin]
      summary: List all platform commands (admin-configurable)
      security: [{PlatformAdminToken: []}]
      responses:
        "200":
          description: Command list
          content:
            application/json:
              schema:
                type: array
                items: {$ref: "#/components/schemas/PlatformCommand"}
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/admin/commands/{command_id}:
    put:
      tags: [admin]
      summary: Update a platform command (enable/disable, reclassify, edit help)
      security: [{PlatformAdminToken: []}]
      parameters:
        - name: command_id
          in: path
          required: true
          schema: {type: integer, format: int64}
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: "#/components/schemas/UpdateCommandRequest"}
      responses:
        "200":
          description: Updated command
          content:
            application/json:
              schema: {$ref: "#/components/schemas/PlatformCommand"}
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/admin/sophub/binding:
    get:
      operationId: adminGetSophubBinding
      summary: Read write-only Sophub binding status
      tags: [admin, sop]
      responses:
        "200":
          description: Binding status without the API key
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SophubBindingStatus"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "500": {$ref: "#/components/responses/InternalError"}
    put:
      operationId: adminBindSophub
      summary: Verify and replace the encrypted Sophub API key
      tags: [admin, sop]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [api_key]
              properties:
                api_key: {type: string, writeOnly: true, minLength: 1}
      responses:
        "200":
          description: Verified binding status; the key is never returned
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SophubBindingStatus"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /v1/admin/sophub/search:
    get:
      operationId: adminSearchSophubSOPs
      summary: Search approved SOP metadata at the fixed Sophub origin
      tags: [admin, sop]
      parameters:
        - {name: q, in: query, required: false, schema: {type: string}}
        - {name: page, in: query, required: false, schema: {type: integer, minimum: 1, default: 1}}
        - {name: page_size, in: query, required: false, schema: {type: integer, minimum: 1, maximum: 100, default: 24}}
      responses:
        "200":
          description: Sophub search result
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SophubSearchResult"}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthorized"}
  # Worker Sophub proxy(方案 §5.2): Runner 经 Platform 受控代理搜索/安装,
  # 携带 ga-sophub-proxy audience 的 per-task capability(不暴露 Sophub Key)。
  /v1/worker/sophub/search:
    get:
      operationId: workerSophubSearch
      summary: Search installable SOPs on behalf of a Runner
      tags: [worker, sop]
      security:
        - BearerToken: []
      parameters:
        - {name: q, in: query, required: true, schema: {type: string, minLength: 1}}
      responses:
        "200":
          description: Installable SOP search result
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SophubSearchResult"}
        "401": {$ref: "#/components/responses/Unauthorized"}
  /v1/worker/sophub/install:
    get:
      operationId: workerSophubInstall
      summary: Fetch one installable SOP on behalf of a Runner
      tags: [worker, sop]
      security:
        - BearerToken: []
      parameters:
        - {name: id, in: query, required: true, schema: {type: string, minLength: 1}}
      responses:
        "200":
          description: Fetch result
          content:
            application/json:
              schema: {$ref: "#/components/schemas/SophubInstallResult"}
        "401": {$ref: "#/components/responses/Unauthorized"}
  # 内部 listener 专用路由(不进主 listener): Runner 经 Platform 受控 MCP
  # proxy 访问外部 MCP Server(server_id → 目标映射即白名单, capability 鉴权)。
...
  /v1/admin/dashboard/stats:
    get:
      security:
        - PlatformAdminToken: []
      operationId: adminDashboardStats
      summary: 仪表盘统计
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {type: object}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "500": {$ref: "#/components/responses/InternalError"}
...
  /v1/admin/invite-codes/{code}:
    delete:
      security:
        - PlatformAdminToken: []
      operationId: adminRevokeInviteCode
      summary: 吊销单个邀请码
      parameters:
        - name: code
          in: path
          required: true
          description: 邀请码
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema: {type: object}
        "400": {$ref: "#/components/responses/ValidationError"}
        "401": {$ref: "#/components/responses/Unauthorized"}
        "500": {$ref: "#/components/responses/InternalError"}
...
components:
  securitySchemes:
    PlatformAdminToken:
      type: apiKey
      in: header
      name: X-Platform-Admin-Token
      description: Required foundation-only loopback development token
    BearerToken:
      type: http
      scheme: bearer
      bearerFormat: platform-user-token
...
  responses:
    ValidationError:
      description: Client-side validation failure (e.g. prompt too long, unknown source)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Missing or invalid X-Platform-Admin-Token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
...
    CreateUserRequest:
      type: object
      required:
        - username
      properties:
        username:
          type: string
          description: Platform username (1-64 bytes)
        password:
          type: string
          description: Optional password; empty means an admin-created test account
    UserRecord:
      type: object
      required:
        - user_id
        - username
        - status
        - created_at
      properties:
        user_id:
          type: integer
          format: int64
        username:
          type: string
        status:
          type: string
          enum: [pending, approved, blocked]
        created_at:
          type: string
          format: date-time
        approved_at:
          type: string
          format: date-time
          nullable: true
...